fix forgotten definition of function in modal_io

+++ src/drivers/actuators/modal_io/modal_io.hpp
@@ -225,6 +225,6 @@ private:
 	int 			parse_response(uint8_t *buf, uint8_t len, bool print_feedback);
 	int			flush_uart_rx();
 	int			check_for_esc_timeout();
-	void			mix_turtle_mode(uint16_t outputs[]);
+	void			mix_turtle_mode(float outputs[]);
 	void			handle_actuator_test();
 };
